move logging and helpers to separate file
