+++ main.go
@@ -1,11 +1,9 @@
 package main
 
 import (
-	"fmt"
 	"log"
 	"net/http"
 	"os"
-	"time"
 )
 
 func main() {
@@ -22,23 +20,4 @@ func main() {
 
 	log.Printf("Running server on port %s...\n", port)
 	http.ListenAndServe(":"+port, LoggingHandler(http.DefaultServeMux))
-}
-
-const LogFormat = `"%s %d %d" %f`
-
-func LoggingHandler(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(res http.ResponseWriter, req *http.Request) {
-
-		startTime := time.Now()
-		next.ServeHTTP(res, req)
-		elapsedTime := time.Now().Sub(startTime)
-
-		requestLine := fmt.Sprintf("%s %s %s", req.Method, req.RequestURI, req.Proto)
-
-		log.Printf(
-			LogFormat,
-			requestLine, http.StatusOK, 0, elapsedTime.Seconds(),
-		)
-
-	})
 }
